move and ignore fee estimate test
