+++ signer/tests/integration/main.rs
@@ -8,6 +8,7 @@ mod communication;
 mod complete_deposit;
 mod contracts;
 mod emily;
+mod fees;
 mod postgres;
 mod rbf;
 mod request_decider;
